requirements.txt
fastapi
uvicorn[standard]
PyJWT
requests
python-dotenv
Flask
msal
fastmcp
azure-identity
